requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
databricks-sdk==0.32.3
datasets==2.21.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
docker==7.1.0
evaluate==0.4.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-auth-oauthlib==1.0.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.66.1
gunicorn==23.0.0
huggingface-hub==0.24.6
hydra-core==1.3.2
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaggle==1.6.17
kiwisolver==1.4.5
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.0.2
mlflow==2.16.2
mlflow-skinny==2.16.2
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==6.0.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.24.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
six==1.16.0
smmap==5.0.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==1.12.1+cu113
torchvision==0.11.2+cu113
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.6
zipp==3.20.1
